Adds initial command-line solitaire game.
This can be played by a human by running 'node main.js' by using
some currently undocumented commands for drawing cards, moving
cards around, adding to the foundation etc. This has been useful
for testing and I'm confident the game code for shuffling,
setting up a game, applying moves and rejecting invalid moves works.
Of course further, automated tests would be an improvement. The
next step is to add a solver since this is not intended to be
played by a human for this project.