Adds split up/down tableau and updates win condition.
The tableau is now an array of {faceUp: [...], faceDown: [...]}
objects, so that possible moves can be calculated easier (no point
in doing move operations of face down cards). Can also isolate
stacks of face up cards easier to cache and prevent recreation etc.

The win condition now just checks that all cards are gone from the
hand and waste and all cards are face up. Any game like this is
definitely solvable and the game doesn't need to go further trying
to solve and possibly backtracking.