Rewrites solver.
Split up getWinningMoves() into a corecursive maybeApplyMove()
function so that we can keep track of which card stacks we've
added to the tableau and then remove them when backtracking.
Also makes the code easier to understand since that recursive
function is not quite so long and doesn't have short circuiting
both within the applyMove() loop and at the beginning of the function.

Two remaining problems: the heuristic is not very good, and the
search space is very large, so the time required to compute a solution
can be very large. A better heuristic will improve this, but if I
want to bound the time I would have to avoid searching the whole solution
space, in which case my solver would only be able to tell if (a) a solution
exists or (b) a solution may or may not exist, which is not satisfying.

The other problem is that because we are searching such a large
solution space, the state cache can become quite large (in the millions
of entries) and memory usage becomes unbounded. To improve on this I
will be switching to an LRU cache in the next diff.