Modifies main.js to read games from stdin.
Games are read in 104-character lines which represent
a shuffled deck (2 characters per card). The line is parsed,
split up into cards and attempted to be solved. diagnostic
info is now written to stderr (things like the current game
state, moves/sec, other unstructured string data). When a
game is won/lost/timeouts a JSON blob is written to stdout
with the winning moves, time elapsed, deck used, etc. It
is written in the same format as the C++ implementation so
that these can be easily compared.

Additionally some bugs have been fixed, and the move type has
been changed to an integer to match the C++ implementation.